Default to NULL for us.format argument
* `ParseEnteredData` and `ParseAsDataFrame` now
set the "us.format" argument to `NULL` by default, so date
strings are checked for both u.s. format and international format
by default.

* Fixes unit test failures in test-dataentry.R that for tests that
started to fail as of flipTime v2.5.0. Dates in international format
(as in the unit tests) should not parse if a user sets us.format to
TRUE (the old default in `ParseEnteredData`).  That they parsed before
v2.5.0 of flipTime was a bug in flipTime.

# How should be self-explanatory with diff
#
Addresses: DS-1685
#
# 50-character subject line
#
# 72-character wrapped longer description.
#
# Don't forget to add reference to Jira issue