DS-3136 Improve unit tests and remove verbs dep [revdep skip]
Unit tests for AsNumeric used verbs::First only for what seems like
an interactive viewing of the output since no expectation
statements were used on their outputs. The unit tests in this
situation have been improved by comparing the structure of the
output to the expected structure after the AsNumeric transform
The occurences of First have been removed and the dep of verbs too.